+++ sprockets/http/runner.py
@@ -8,6 +8,7 @@ Run a Tornado HTTP service.
 """
 import logging
 import signal
+import sys
 
 from tornado import httpserver, ioloop
 
@@ -47,8 +48,12 @@ class Runner(object):
         self.shutdown_limit = 5
         try:
             self.application.runner_callbacks.setdefault('shutdown', [])
+            self.application.runner_callbacks.setdefault('before_run', [])
         except AttributeError:
-            setattr(self.application, 'runner_callbacks', {'shutdown': []})
+            setattr(self.application, 'runner_callbacks', {
+                'shutdown': [],
+                'before_run': [],
+            })
 
     def start_server(self, port_number, number_of_procs=0):
         """
@@ -87,11 +92,24 @@ class Runner(object):
 
         If the application's ``debug`` setting is ``True``, then we are
         going to run in a single-process mode; otherwise, we'll let
-        tornado decide how many sub-processes to spawn.
+        tornado decide how many sub-processes to spawn.  In any case, the
+        applications *before_run* callbacks are invoked.  If a callback
+        raises an exception, then the application is terminated by calling
+        :func:`sys.exit`.
 
         """
+        iol = ioloop.IOLoop.instance()
         self.start_server(port_number, number_of_procs)
-        ioloop.IOLoop.instance().start()
+        for callback in self.application.runner_callbacks['before_run']:
+            try:
+                callback(self.application, iol)
+            except Exception:
+                self.logger.error('before_run callback %r cancelled start',
+                                  callback, exc_info=1)
+                self._shutdown()
+                sys.exit(70)
+
+        iol.start()
 
     def _on_signal(self, signo, frame):
         self.logger.info('signal %s received, stopping', signo)